+++ src/ScottPlot5/ScottPlot5/Primitives/ConnectStyle.cs
@@ -1,0 +1,19 @@
+﻿namespace ScottPlot;
+
+public enum ConnectStyle
+{
+    /// <summary>
+    /// Connect points with straight lines
+    /// </summary>
+    Straight,
+
+    /// <summary>
+    /// Connect points with a line horizontally, then vertically
+    /// </summary>
+    StepHorizontal,
+
+    /// <summary>
+    /// Connect points with a line vertically, then horizontally
+    /// </summary>
+    StepVertical,
+}