+++ CHANGELOG.md
@@ -10,6 +10,7 @@ _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-01-18_
 * Axes: Improved automatic axis determination for plots containing non-real or infinite data limits (#3232, #3237)
 * Bar Plots: Do not overwrite existing colors when adding `Bar` collections to the plot (#3231)
 * Label: Clear cached typefaces when styles change (#3236) _Thanks @kl7107 and @prime167_
+* Scatter Plot: Added a `ConnectStyle` to enable step display mode (#3242) _Thanks @NoahSigl_
 
 ## ScottPlot 5.0.18
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-01-17_